module_backend/requirements.txt
# --- Core FastAPI stack ---
fastapi==0.115.12
uvicorn[standard]==0.34.2
starlette==0.46.2

# --- Database & ORM ---
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10

# --- Authentication & Security ---
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose==3.4.0
cryptography==45.0.2
ecdsa==0.19.1
rsa==4.9.1

# --- Validation & Settings ---
pydantic==2.11.4
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-multipart==0.0.20
email_validator==2.2.0

# --- Utilities ---
requests==2.32.3
aiofiles==24.1.0
httpx==0.28.1
watchfiles==1.0.5

#email
sendgrid==6.12.5

# --- Optional CLI / Pretty logging ---
#typer==0.15.3
#rich==14.0.0
#colorama==0.4.6